Rename test/test-data/samplers to test/test-data/sampler
