+++ tests/test-data/sampler/gamma.js
